+++ spec/truffle/interop/special_forms_spec.rb
@@ -55,6 +55,12 @@ describe "Interop special form" do
     @object.log.should include("KEYS")
   end
 
+  guard -> { !Truffle.native? } do
+    it "#class sends READ('class') on Java class objects" do
+      Java.type('java.math.BigInteger').class.getName.should == 'java.math.BigInteger'
+    end
+  end
+
   it "#name sends INVOKE" do
     @object.foo
     @object.bar(1, 2, 3)